Commented out magic library
Mapped all filters to the mapping dictionary

Added email filter

Divided filters in image and document filters
