+++ tim/Tester.py
@@ -6,8 +6,6 @@ Created on Jan 19, 2016
 import logging
 from tim.metadataExtractor.Extractor import Extractor
 
-
-
 # log lvls: DEBUG - INFO - WARNING - ERROR - CRITICAL
 logging.basicConfig(level=logging.ERROR, format='%(asctime)s - %(module)s - %(levelname)s ##\t  %(message)s')
 
@@ -16,14 +14,10 @@ thisContainer = 'filter_test'
 log = logging.getLogger()
 ex = Extractor(containerName=thisContainer)
 
-
-
-
 if __name__ == '__main__':
-	log.error('starting metadata extractor tester')
-	
-	
-	ex.runIdentifierForWholeContainer()
-	ex.runFilterForWholeContainer()
-	
-	log.error('DONE')
+    log.error('starting metadata extractor tester')
+
+    ex.runIdentifierForWholeContainer()
+    ex.runFilterForWholeContainer()
+
+    log.error('DONE')
